sensors: Create bool to select what timestamp to use
Older devices may have an issue with rotation freezes up and
requires a reboot to fix. In deep sleep the sensor's timestamp
is far off, depending how long it's in sleep, causing rotation
not to work. onSensorChanged if true it will use
SystemClock.elapsedRealtimeNanos() instead of event.timestamp.
Possibly an update to the custom sensor libs.

UA include:
 sensors: Rename bool for selecting what timestamp to use for orientation sensor
 Author: Arne Coucheron <arco68@gmail.com>
 Date: Wed Mar 9 20:11:03 2016 +0100

Change-Id: I0b487af370b7a924b2498ec7e07daf24a7b27a09